moved Openstack related functions to a separate class
